cloned old project, pretty much for its user model
The cloned project (CandidateTracking) is a project I'm building to
allow employers seeking employees to post opportunities and allows the
potential candidates (seekers) to both post their resumes and apply to
opportunities.  This project has a (very) good system of managing
multiple users (admin, recruiter and applicant) already built into it;
thus, altering this system for the needs of the "AdProj" is better than
rolling out a whole new authentication/authorization system.

On a side note, this is the first time I'll have not only mined but
cloned one of my previous projects for use in a new project, and I'm
excited and confident that the result will be capable (though possibly
not beautiful).